+++ data-raw/Parvin73/Parvin73.R
@@ -18,3 +18,12 @@ Parvin73 %>% select(-note) %>%
   rename(d_pci = incpcgrowth) -> Parvin73
 
 save(Parvin73, file="data/Parvin73.rda")
+
+Parvin73 %>%
+  mutate(pop = ifelse(country == "Belgium", 91.84, pop),
+         ln_levviol = log(levviol + .001)) %>%
+  log_at(c("pci", "incdist", "d_pci", "sem", "comint", "concfac")) -> Data
+
+summary(M1 <- lm(ln_levviol ~ ln_pci + ln_incdist + ln_d_pci + ln_sem +
+                   ln_comint + ln_concfac,
+                 data = Data))