+++ NEWS.md
@@ -1,3 +1,11 @@
+# stevedata 1.5.0
+
+Additions include:
+
+- `Russett64`: A data set on inequality and instability, to replicate Russett's (1964) analysis.
+- `china_peace`: A data set on Chinese peace engagement in conflict-affected countries. These are useful for teaching about panel models, fixed effects, or some GLMs like the negative binomial model.
+- `wb_groups`: A data set on World Bank country groups/classifications, for ease of identifying just the three character ISO codes of interest for use in the `{WDI}` package.
+
 # stevedata 1.4.0
 
 Additions include: